+++ src/main/java/es/in2/vcverifier/service/JWTService.java
@@ -21,5 +21,5 @@ public interface JWTService {
 
     Object getVCFromPayload(Payload payload);
 
-
+    String generateJWTwithOI4VPType(String s);
 }